Update ConfigureReporting from Module to Class
